Move the libs/pubsub package to internal scope (#7451)
No API changes, merely changes the import path.
