+++ internal/eventbus/event_bus.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 
 	abci "github.com/tendermint/tendermint/abci/types"
+	tmpubsub "github.com/tendermint/tendermint/internal/pubsub"
 	"github.com/tendermint/tendermint/libs/log"
-	tmpubsub "github.com/tendermint/tendermint/libs/pubsub"
 	"github.com/tendermint/tendermint/libs/service"
 	"github.com/tendermint/tendermint/types"
 )
