+++ internal/pubsub/query/query.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"github.com/tendermint/tendermint/abci/types"
-	"github.com/tendermint/tendermint/libs/pubsub/query/syntax"
+	"github.com/tendermint/tendermint/internal/pubsub/query/syntax"
 )
 
 // All is a query that matches all events.
