+++ internal/pubsub/query/syntax/doc.go
