+++ internal/pubsub/query/syntax/scanner.go
