+++ internal/state/indexer/block/kv/util.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	"github.com/google/orderedcode"
-	"github.com/tendermint/tendermint/libs/pubsub/query/syntax"
+	"github.com/tendermint/tendermint/internal/pubsub/query/syntax"
 	"github.com/tendermint/tendermint/types"
 )
