+++ internal/state/indexer/block/null/null.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"errors"
 
+	"github.com/tendermint/tendermint/internal/pubsub/query"
 	"github.com/tendermint/tendermint/internal/state/indexer"
-	"github.com/tendermint/tendermint/libs/pubsub/query"
 	"github.com/tendermint/tendermint/types"
 )
