+++ internal/state/indexer/eventsink.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	abci "github.com/tendermint/tendermint/abci/types"
-	"github.com/tendermint/tendermint/libs/pubsub/query"
+	"github.com/tendermint/tendermint/internal/pubsub/query"
 	"github.com/tendermint/tendermint/types"
 )
