+++ internal/state/indexer/indexer.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 
 	abci "github.com/tendermint/tendermint/abci/types"
-	"github.com/tendermint/tendermint/libs/pubsub/query"
+	"github.com/tendermint/tendermint/internal/pubsub/query"
 	"github.com/tendermint/tendermint/types"
 )
