+++ internal/state/indexer/tx/null/null.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 
 	abci "github.com/tendermint/tendermint/abci/types"
+	"github.com/tendermint/tendermint/internal/pubsub/query"
 	"github.com/tendermint/tendermint/internal/state/indexer"
-	"github.com/tendermint/tendermint/libs/pubsub/query"
 )
 
 var _ indexer.TxIndexer = (*TxIndex)(nil)
