+++ rpc/client/http/ws.go
@@ -8,8 +8,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/tendermint/tendermint/internal/pubsub"
 	tmjson "github.com/tendermint/tendermint/libs/json"
-	"github.com/tendermint/tendermint/libs/pubsub"
 	rpcclient "github.com/tendermint/tendermint/rpc/client"
 	"github.com/tendermint/tendermint/rpc/coretypes"
 	jsonrpcclient "github.com/tendermint/tendermint/rpc/jsonrpc/client"
